[mintscan] remove utils package and move Respond to models
