+++ mintscan/models/rest.go
@@ -1,4 +1,4 @@
-package utils
+package models
 
 import (
 	"encoding/json"
